Finish App Localization✅
- Added `easy_localization` package to `pubspec.yaml`.
- Finished localization setup within `runApp` and `MaterialApp`.
- Added all needed strings to the `app_strings` file.
- Added `ar.json` and `en.json` for English and Arabic languages.
- Refactor back icons and `SearchResultCard` to be appropriated with the new language direction.
- Refactor all text widgets to display the new language strings.
- Refactor the `WeatherRequestBody` and `CityWeatherRequestBody` to send the language code within the requests.
- Refactor `HomeError` to receive `Failure` instead of `String` message.

+++ lib/features/draggable_sheet/logic/cubit/sheet_cubit.dart
@@ -56,8 +56,8 @@ class SheetCubit extends Cubit<SheetState> {
         final dateKey = forecastedWeather[index].keys.first;
         return WeeklyConditionItemModel(
           date: dateKey,
-          highestTemp: forecastedWeather[index][dateKey]!.last,
-          lowestTemp: forecastedWeather[index][dateKey]!.first,
+          highestTemp: forecastedWeather[index][dateKey]!.last.toDouble(),
+          lowestTemp: forecastedWeather[index][dateKey]!.first.toDouble(),
         );
       },
     );
